Use blobstore for images
Previously images and icons were stored in the datastore as properties
on the recipe. Now images are stored in the blobstore and referenced to
in the Recipe model. The image server can handle resizing for icons,
etc..

Includes changes to the service to pass back a URL for the image in a
get recipe call